+++ app.ts
@@ -1,8 +1,7 @@
 import { serve } from "https://deno.land/std@0.170.0/http/server.ts";
 import { Hono } from "https://deno.land/x/hono@v2.7.0/mod.ts";
-import { deployment, location } from "./app/deploy.ts";
+import { deployment, location } from "./app/deployments.ts";
 import { handler } from "./app/webhooks.ts";
-import { verifyRequest } from "./app/qstash.ts";
 
 const app = new Hono();
 
@@ -25,18 +24,6 @@ app.get("/", (context) => context.text("Hello, I'm Denopendabot!"));
 app.post("/api/github/webhooks", async (context) => {
   await handler(context.req);
   return context.json(null, 200);
-});
-
-// verify requests from qstash
-app.use("/api/qstash/*", async (c, next) => {
-  const valid = await verifyRequest({
-    signature: c.req.header("upstash-signature"),
-    body: await c.req.text(),
-  });
-  if (!valid) {
-    throw new Error("Signature is invalid");
-  }
-  await next();
 });
 
 await serve(app.fetch, { onListen: () => {} });
