+++ app/env.ts
@@ -1,4 +1,4 @@
-import { configSync } from "https://deno.land/std@0.170.0/dotenv/mod.ts";
+import { loadSync } from "https://deno.land/std@0.170.0/dotenv/mod.ts";
 
 type Env = {
   CI?: string;
@@ -17,17 +17,13 @@ type Env = {
   CLIENT_ID: string;
   CLIENT_SECRET: string;
   WEBHOOK_SECRET: string;
-
-  UPSTASH_REDIS_REST_URL: string;
-  UPSTASH_REDIS_REST_TOKEN: string;
-  QSTASH_TOKEN: string;
 };
 
 const getEnv = () => {
   if (Deno.env.get("CI") || Deno.env.get("DENO_DEPLOYMENT_ID")) {
     return Deno.env.toObject() as Env;
   } else {
-    return configSync({ export: true }) as Env;
+    return loadSync({ export: true }) as Env;
   }
 };
